model/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.35
marshmallow==3.21.3
Werkzeug==3.0.4
itsdangerous==2.2.0
PyJWT==2.9.0
passlib==1.7.4
pytest==8.3.3
requests==2.32.3
